Use poolstl to sort randstrobes in parallel
See https://github.com/alugowski/poolSTL

Sorting the randstrobes is currently a bottleneck in index generation as it
does not run in parallel. This is an attempt to parallelize it.

poolstl’s sort uses regular std::sort under the hood.
We currently use pdqsort_branchless, which is about twice as fast as
std::sort, so parallel sorting breaks even with pdqsort_branchless at about
2-3 threads. It gets faster with more threads, but not as much as one would
perhaps expect. Here are the sorting runtimes for CHM13:

- 31 s with pdqsort_branchless
- 59 s with std::sort
- 34 s with parallel sort, 2 threads
- 24 s with parallel sort, 4 threads
- 23 s with parallel sort, 8 threads

Another issue is that sorting is no longer in place, so memory usage goes
up by a couple of gigabytes, which is another reason for me not to make this
change.

+++ src/index.hpp
@@ -51,7 +51,7 @@ struct StrobemerIndex {
 
     void write(const std::string& filename) const;
     void read(const std::string& filename);
-    void populate(float f, size_t n_threads);
+    void populate(float f, unsigned n_threads);
     void print_diagnostics(const std::string& logfile_name, int k) const;
     int pick_bits(size_t size) const;
     size_t find(randstrobe_hash_t key) const {
